Contact info

+++ controllers/userController.js
@@ -1,39 +1,45 @@
+
 import Users from "../models/userModel.js";
 
 export const login = async (req, res) => {
   const { email, password } = req.body;
-try {
-  
-  if (!email || !password)
-    return res
-      .status(500)
-      .json({ success: false, error: "Enter A Correct Information" });
+  try {
+    if (!email || !password)
+      return res
+        .status(500)
+        .json({ success: false, error: "Enter A Correct Information" });
 
-  const user = await Users.findOne({ email }).select("+password");
-  if (!user) {
-    return res
-      .status(500)
-      .json({ success: false, error: "Inavalid Credentials: Email" });
+    const user = await Users.findOne({ email }).select("+password");
+    if (!user) {
+      return res
+        .status(500)
+        .json({ success: false, error: "Inavalid Credentials: Email" });
+    }
+    console.log(user);
+
+    const isCorrect = await user.comparePassword(password);
+    if (!isCorrect) {
+      return res
+        .status(500)
+        .json({ success: false, error: "Inavalid Credentials: Password" });
+    }
+
+    const token = user.createJWT();
+    user.password = undefined;
+    // res.status(200).json({ success: true, user: user });
+    res
+      .status(200)
+      .header("Authorization", "Bearer " + token)
+      .json({ success: true, user: user, token: token });
+  } catch (error) {
+    res.status(500).json({ success: false, error: error });
   }
-  console.log(user);
-
-  const isCorrect = await user.comparePassword(password);
-  if(!isCorrect){ return res.status(500).json({ success: false, error: "Inavalid Credentials: Password"}); }
-
-  const token = user.createJWT();
-  user.password= undefined;
-  // res.status(200).json({ success: true, user: user });
-res.status(200).header("Authorization", "Bearer " + token).json({success: true, user: user, token: token});
-
-} catch (error) {
-  res.status(500).json({success: false, error: error});
-}
 };
 
 /**
- * Create a new user 
- * @param {*} req 
- * @param {*} res 
+ * Create a new user
+ * @param {*} req
+ * @param {*} res
  */
 export const createUser = async (req, res) => {
   const { name, email, password, IsAdmin } = req.body;
@@ -100,6 +106,9 @@ export const getUsers = async (req, res) => {
  */
 export const deleteUser = async (req, res) => {
   try {
+
+    const total = await Users.countDocuments({});
+if (total === 1 ){ return res.status(405).json({success:false , error:'Cann\'t Delet yourself '});}
     const user = await Users.findByIdAndDelete(req.params.id);
     if (!user) {
       return res.status(404).json({
@@ -107,9 +116,41 @@ export const deleteUser = async (req, res) => {
         error: "User not found",
       });
     }
+console.log(total);
     res.status(200).json({
       success: true,
       error: "User deleted successfully",
+    });
+  } catch (error) {
+    console.error(error);
+    res.status(500).json({
+      success: false,
+      error: "Server error",
+    });
+  }
+};
+
+/**
+ * Delete Users By select the id of the user
+ * @param {*} req 
+ * @param {*} res 
+ * @returns 
+ */
+export const deleteUsers = async (req, res) => {
+  try {
+    const { ids } = req.body;
+    const result = await Users.deleteMany({ _id: { $in: ids } });
+    if (result.deletedCount === 0) {
+      return res.status(404).json({
+        success: false,
+        error: "Users not found",
+      });
+    }
+    const total = await Users.countDocuments({});
+    console.log(total);
+    res.status(200).json({
+      success: true,
+      message: `Deleted ${result.deletedCount} user(s) successfully`,
     });
   } catch (error) {
     console.error(error);
@@ -134,19 +175,18 @@ export const updateUser = async (req, res) => {
         error: "User not found",
       });
     }
-    res.status(200).json({
+    return res.status(200).json({
       success: true,
       data: updatedUser,
     });
   } catch (error) {
     console.error(error);
-    res.status(500).json({
+    return res.status(500).json({
       success: false,
-      error: error
+      error: error,
     });
   }
 };
-
 
 /**
  * Function to read a  user by id from the database
@@ -154,23 +194,23 @@ export const updateUser = async (req, res) => {
  * @param {*} res
  * @returns
  */
-export const getUser = async (req, res ) => {
+export const getUser = async (req, res) => {
   const id = req.params.id;
   try {
-    const user = await Users.find({_id:id});
+    const user = await Users.find({ _id: id });
     if (!user) {
       return res.status(404).json({
         success: false,
         error: "No users found",
       });
     }
-    res.status(200).json({
+    return res.status(200).json({
       success: true,
       data: user,
     });
   } catch (error) {
     console.error(error);
-    res.status(500).json({
+    return res.status(500).json({
       success: false,
       error: "Server error",
     });
@@ -183,25 +223,34 @@ export const getUser = async (req, res ) => {
  * @param {*} res
  * @returns
  */
-export const getUserbyName = async (req, res ) => {
+export const getUserbyName = async (req, res) => {
   const name = req.query.name;
   try {
-    const user = await Users.find({name: name});
+    const user = await Users.find({ name: { $regex: `.*${name}.*`, $options: 'i' } });
     if (!user) {
       return res.status(404).json({
         success: false,
         error: "No users found",
       });
     }
-    res.status(200).json({
+    return res.status(200).json({
       success: true,
       data: user,
     });
   } catch (error) {
     console.error(error);
-    res.status(500).json({
+    return res.status(500).json({
       success: false,
       error: "Server error",
     });
   }
 };
+
+
+
+
+
+
+
+
+
